requirements.txt
python-dateutil
six
exifread
